[#380]; refactor: DevPrivilegeTestHolder를 support/holder로 이동
- Holder: DB 없이 메모리에 상태를 들고 있다가 다른 컴포넌트가 읽고 쓰게 해주는 컴포넌트(스프링 SecurityContextHolder와 같은 패턴). 파일이 하나뿐이어도 exception/converter처럼 기능 폴더로 묶기 위해 분리
